+++ packages/syntax/src/lexer.ts
@@ -59,7 +59,6 @@ const PUNCTUATION = [
     '@',
     "'",
     '~',
-    '×',
 ] as const;
 
 const isDigit = (c: string) => c >= '0' && c <= '9';