+++ packages/syntax/src/print.ts
@@ -685,7 +685,6 @@ class Printer {
                     this.expression(right, 6, inner),
                 ];
             case '*':
-            case '×':
             case '/':
                 return [
                     opensRight(left) ? this.parenthesized(left) : this.expression(left, 6, context),